+++ cmd/mr-orchestrate/main.go
@@ -10,7 +10,7 @@ import (
 	"os"
 )
 
-var version = "0.36.2"
+var version = "0.37.0"
 
 func main() {
 	if len(os.Args) < 2 {